% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQtxt.R
\name{getInfo}
\alias{getInfo}
\title{Pulls Attribute info}
\usage{
getInfo(var, attr = c("questionText", "generalInfo"))
}
\arguments{
\item{var}{A single variable with attributes}

\item{attr}{A vector of attributes to pull}
}
\value{
The info contained in the attribute/s
}
\description{
This function is a helper that pulls attribute information
}
