% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDataFrameList.R
\name{mergeDataframeList}
\alias{mergeDataframeList}
\title{Custom merge wrapper}
\usage{
mergeDataframeList(dataFrameList, matchingKey)
}
\arguments{
\item{dataFrameList}{A list of dataFrames to be merged into a single dataFrame}

\item{matchingKey}{The unique key to be used to merge the dataFrames}
}
\value{
A merged dataFrame
}
\description{
This function merges a list of dataFrames by a common key variable
using base merge and purrr::reduce
}
