% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeVariable.R
\name{recode_variable}
\alias{recode_variable}
\title{Recode a variable}
\usage{
recode_variable(variable, presort = TRUE)
}
\arguments{
\item{variable}{A variable to be recoded 1-n}
}
\value{
A recoded variable
}
\description{
Takes a variable and returns a numeric recoded version
from 1-n unique values
}
