% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeVariable.R
\name{recodeVariable}
\alias{recodeVariable}
\title{Recode a variable}
\usage{
recodeVariable(var)
}
\arguments{
\item{var}{A variable to be recoded 1-n}
}
\value{
A recoded variable
}
\description{
Takes a variable and returns a numeric recoded version
from 1-n unique values
}
